Add -version switch to the compiler to show the thrift version [THRIFT-40]
The version is determined by a toplevel VERSION file. After this version
is shown the revision. For subversion, it is simply the current revision
as given by `svnversion`. For git, the latest svn revision that HEAD includes
is shown, followed by the offset of HEAD from that commit, followed by a
truncated sha1 for HEAD. If the offset is 0, the offset and sha1 are omitted.